+++ frontend/src/App.css
@@ -100,3 +100,55 @@ button {
   background-color: #e1bee7;
 }
 
+.add-button {
+  margin-top: 20px;
+  padding: 10px;
+  border-radius: 5px;
+  background-color: #6a1b9a;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.add-button:hover {
+  background-color: #8e24aa;
+}
+
+.add-form {
+  margin-top: 20px;
+  padding: 20px;
+  background-color: rgba(255, 255, 255, 0.8);
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
+
+.add-form label {
+  display: block;
+  margin-bottom: 10px;
+  font-size: 1rem;
+}
+
+.add-form input,
+.add-form select {
+  padding: 8px;
+  width: 100%;
+  margin-top: 5px;
+  border-radius: 4px;
+  border: 1px solid #ccc;
+}
+
+.add-form button {
+  padding: 10px 20px;
+  margin-top: 10px;
+  background-color: #7b1fa2;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.add-form button:hover {
+  background-color: #9c27b0;
+}
+
+
